requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==21.10b0
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==4.8.0
contourpy==1.2.0
cryptography==42.0.3
cycler==0.12.1
dacite==1.8.1
dataclasses-json==0.6.3
datasets==2.16.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
flake8==4.0.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gradio==4.15.0
gradio_client==0.8.1
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipython==8.20.0
isort==5.10.1
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.3
langchain-community==0.0.16
langchain-core==0.1.16
langchain-openai==0.0.5
langsmith==0.0.83
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy==0.910
mypy_extensions==0.4.4
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.16.3
openai==1.11.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.12
overrides==7.6.0
packaging==23.2
pandas==2.2.0
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
posthog==3.3.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
pyflakes==2.4.0
Pygments==2.17.2
PyNaCl==1.5.0
pyparsing==3.1.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pysen==0.10.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruff==0.1.14
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sshtunnel==0.4.0
stack-data==0.6.3
starlette==0.35.1
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.1
toml==0.10.2
tomli==1.2.3
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.1
traitlets==5.14.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unidiff==0.7.5
urllib3==2.1.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==11.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
